adp getVol...() fix

+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.adp/src/com/mmxlabs/models/lng/adp/DistributionModel.java
@@ -143,7 +143,7 @@ public interface DistributionModel extends MMXObject {
 	 * @model kind="operation"
 	 * @generated
 	 */
-	int getModelOrContractVolumePerCargo();
+	double getModelOrContractVolumePerCargo();
 
 	/**
 	 * <!-- begin-user-doc -->
